+++ src/algorithms/signal_source/adapters/four_bit_cpx_file_signal_source.cc
@@ -69,7 +69,7 @@ FourBitCpxFileSignalSource::FourBitCpxFileSignalSource(
 std::tuple<size_t, bool> FourBitCpxFileSignalSource::itemTypeToSize()
 {
     auto is_complex = false;
-    auto item_size = size_t(sizeof(char));  // default
+    auto item_size = sizeof(char);  // default
 
     if (item_type() == "byte")
         {
